Moved sisphus to separate role
I've moved sisyphus to a separate role so that it can be installed to
testuppmax. I also made a ugly trick - adding a small script called
sbatch the path which only echos anything passed to it to "fake" a slurm
environment. This is probably something that we want to do better in the
future.